Move App and Routes into components dir
